+++ react/components/SearchResultContainer.js
@@ -1,44 +1,12 @@
-import React, { useState, useRef, useCallback } from 'react'
-import { min } from 'ramda'
-
-import { Container } from 'vtex.store-components'
+import React from 'react'
+import InfiniteScroll from 'react-infinite-scroll-component'
 
 import { PopupProvider } from './Popup'
-import InfiniteScrollLoaderResult from './loaders/InfiniteScrollLoaderResult'
-import ShowMoreLoaderResult from './loaders/ShowMoreLoaderResult'
+import SearchResult from './SearchResult'
 import { searchResultContainerPropTypes } from '../constants/propTypes'
-import {
-  useSearchPageStateDispatch,
-  useSearchPageState,
-} from 'vtex.search-page-context/SearchPageContext'
-
-const PAGINATION_TYPES = ['show-more', 'infinite-scroll']
-
-const useFetchingMore = () => {
-  const [fetchMoreLoading, localSetMore] = useState(false)
-  const { isFetchingMore } = useSearchPageState()
-  const dispatch = useSearchPageStateDispatch()
-  const setFetchMore = useCallback(
-    value => {
-      dispatch({ type: 'SET_FETCHING_MORE', args: { isFetchingMore: value } })
-      localSetMore(value)
-    },
-    [dispatch]
-  )
-  const stateValue =
-    isFetchingMore === undefined ? fetchMoreLoading : isFetchingMore
-  return [stateValue, setFetchMore]
-}
-
-const useFetchMoreOnStateChange = (fetchMore, isFetchingMore) => {
-  const isFetchingRef = useRef(false)
-  const isFetchingPrevious = isFetchingRef.current
-  // Fire fetch more if user just pressed on button, save previous state on ref to not get lost
-  if (isFetchingMore && !isFetchingPrevious) {
-    fetchMore()
-  }
-  isFetchingRef.current = isFetchingMore
-}
+import { useFetchMore } from '../hooks/useFetchMore'
+import { PAGINATION_TYPE } from '../constants/paginationType'
+import { Container } from 'vtex.store-components'
 
 /**
  * Search Result Container Component.
@@ -61,87 +29,58 @@ const SearchResultContainer = props => {
       variables: { query },
     },
     pagination,
+    page,
     children,
   } = props
 
-  const [fetchMoreLoading, setFetchMoreLoading] = useFetchingMore()
+  const {
+    handleFetchMoreNext,
+    handleFetchMorePrevious,
+    loading: fetchMoreLoading,
+    from,
+    to,
+  } = useFetchMore(page, recordsFiltered, maxItemsPerPage, fetchMore, products)
 
-  const fetchMoreLocked = useRef(false)
+  const resultComponent = children || (
+    <SearchResult
+      {...props}
+      breadcrumbsProps={{ breadcrumb }}
+      onFetchMore={handleFetchMoreNext}
+      fetchMoreLoading={fetchMoreLoading}
+      onFetchPrevious={handleFetchMorePrevious}
+      pagination={pagination}
+      query={query}
+      loading={loading}
+      recordsFiltered={recordsFiltered}
+      products={products}
+      brands={brands}
+      specificationFilters={specificationFilters}
+      priceRanges={priceRanges}
+      tree={categoriesTrees}
+      to={to}
+      from={from}
+    />
+  )
 
-  const handleFetchMore = () => {
-    if (fetchMoreLocked.current || products.length === 0) {
-      return
-    }
+  const infiniteScrollComponent = (
+    <InfiniteScroll
+      style={{ overflow: 'none' }}
+      dataLength={products.length}
+      next={handleFetchMoreNext}
+      hasMore={to + 1 < recordsFiltered}
+      useWindow={false}
+    >
+      {resultComponent}
+    </InfiniteScroll>
+  )
 
-    fetchMoreLocked.current = true
-
-    const to = min(maxItemsPerPage + products.length, recordsFiltered) - 1
-
-    setFetchMoreLoading(true)
-
-    fetchMore({
-      variables: {
-        from: products.length,
-        to,
-      },
-      updateQuery: (prevResult, { fetchMoreResult }) => {
-        setFetchMoreLoading(false)
-        fetchMoreLocked.current = false
-
-        // backwards compatibility
-        if (prevResult.search) {
-          return {
-            search: {
-              ...prevResult.search,
-              products: [
-                ...prevResult.search.products,
-                ...fetchMoreResult.search.products,
-              ],
-            },
-          }
-        }
-
-        return {
-          ...prevResult,
-          productSearch: {
-            ...prevResult.productSearch,
-            products: [
-              ...prevResult.productSearch.products,
-              ...fetchMoreResult.productSearch.products,
-            ],
-          },
-        }
-      },
-    })
-  }
-
-  useFetchMoreOnStateChange(handleFetchMore, fetchMoreLoading)
-
-  const ResultComponent =
-    pagination === PAGINATION_TYPES[0]
-      ? ShowMoreLoaderResult
-      : InfiniteScrollLoaderResult
+  const isInfiniteScroll = pagination != PAGINATION_TYPE.SHOW_MORE
 
   return (
     <Container className="pt3-m pt5-l">
       <PopupProvider>
         <div id="search-result-anchor" />
-        <ResultComponent
-          {...props}
-          breadcrumbsProps={{ breadcrumb }}
-          onFetchMore={handleFetchMore}
-          fetchMoreLoading={fetchMoreLoading}
-          query={query}
-          loading={loading}
-          recordsFiltered={recordsFiltered}
-          products={products}
-          brands={brands}
-          specificationFilters={specificationFilters}
-          priceRanges={priceRanges}
-          tree={categoriesTrees}
-        >
-          {children}
-        </ResultComponent>
+        {isInfiniteScroll ? infiniteScrollComponent : resultComponent}
       </PopupProvider>
     </Container>
   )