+++ react/components/SelectionListItem.js
@@ -4,11 +4,22 @@ import classNames from 'classnames'
 import searchResult from '../searchResult.css'
 
 const SelectionListItem = ({ option, onItemClick, selected }) => {
-  const { setQuery } = useRuntime()
+  const { navigate } = useRuntime()
 
   const handleOptionClick = () => {
     onItemClick()
-    setQuery({ order: option.value })
+    const urlParams = new URLSearchParams(window.location.search)
+    urlParams.set('order', option.value)
+    urlParams.delete('page')
+    /* Ideally this should be a setQuery since it behaves better in terms of
+    UX and performance, but right now this is not possible because the setQuery
+    alone does not reset the values of useFetchMore, which causes a bug on the
+    fetch more/previous buttons behaviour.
+    */
+    navigate({
+      to: window.location.pathname,
+      query: urlParams.toString(),
+    })
   }
 
   const highlight = selected ? 'bg-light-gray' : 'hover-bg-muted-5 bg-base'